LayoutTests/imported/w3c/web-platform-tests/html/canvas/offscreen/path-objects/2d.path.roundrect.nonfinite-expected.txt

2d.path.roundrect.nonfinite

roundRect() with Infinity/NaN is ignored


PASS roundRect() with Infinity/NaN is ignored
